requirements.txt
# Core Identity
spiffe>=0.2.0

# Server & Client
aiohttp
streamlit

# APIs
tavily-python

# Utilities
authlib
langchain_community
langchain_core
cryptography

# Observability (OTEL)
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-aiohttp-server
opentelemetry-instrumentation-aiohttp-client
opentelemetry-instrumentation-logging
